refactor of BOSSSpatialPyramids -> SpatialBOSS
